r717/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-modeltranslation==0.17.1
future==0.18.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
numpy==1.19.4
packaging==20.3
pandas==1.1.4
pep517==0.8.2
Pillow==8.2.0
pkg-resources==0.0.0
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.25.8
viberbot==1.0.12
webencodings==0.5.1
